add skill editor

+++ binedit/binedit.cpp
@@ -638,7 +638,7 @@ static void patch_skills(const Path& data, const Path& json)
         skill.effect_id = node.get<uint16_t>("effect_id");
         skill.effect_target = node.get<uint8_t>("effect_target");
         skill.effect_type = node.get<uint8_t>("ynk_effect_type");
-        skill.ynk_id = node.get<uint8_t>("ynk_id");
+        skill.ynk_id = node.get<uint16_t>("ynk_id");
 
         skill.write(&file[pos]);
     }
